backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-dotenv==1.0.0
firebase-admin==6.6.0
pydantic==2.10.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tensorflow-cpu==2.21.0
numpy==2.1.3
scikit-learn==1.5.2
google-generativeai==0.8.3